ability to play a playlist/your history on it now integrated
