+++ my-app/src/App.css
@@ -1,3 +1,7 @@
 .page {
   padding-top: max(7vh, 3em); /* Ofset for NavBar */
 }
+
+.dark :not(img) {
+  filter: invert(100%);
+}
