+++ my-app/src/components/Card/Card.css
@@ -44,6 +44,6 @@
 }
 
 .card-title {
-    font-size: 1.2em;
-    font-weight: bold;
+    font-size: 1.25em;
+    font-weight: 500;
 }
