+++ my-app/src/components/FlipSide/FlipSide.css
@@ -2,11 +2,16 @@
     grid-column-start: 1;
     grid-row-start: 1;
     -webkit-backface-visibility: hidden;
+    -moz-backface-visibility: hidden;
     backface-visibility: hidden;
+    /* -webkit-transform: rotateX(0deg); */
+    transform: rotateX(0deg);
 }
 
 .back {
+    -webkit-transform: rotateY(180deg);
     transform: rotateY(180deg);
+    z-index: 1;
 }
 
 .wrapped-side {
