+++ my-app/src/components/FlipWrap/FlipWrap.css
@@ -1,9 +1,12 @@
 .flippers {
     display: grid;
+    -webkit-transition: transform 1s;
     transition: transform 1s;
+    -webkit-transform-style: preserve-3d;
     transform-style: preserve-3d;
 }
 
 .flipped {
+    -webkit-transform: rotateY(180deg);
     transform: rotateY( 180deg );
 }
