+++ my-app/src/state/slices/view/viewSlice.js
@@ -1,21 +1,17 @@
 import { createSlice } from '@reduxjs/toolkit';
+import reducers from './viewReducer';
 
 const grid = 'grid';
 const list = 'list';
 
 const initialState = {
     value: grid
-}
+};
 
 export const viewSlice = createSlice({
     name: 'view',
     initialState,
-    reducers: {
-        toggleView: (state) => {
-            const isCurrentlyGrid = state.value === grid;
-            state.value = isCurrentlyGrid ? list : grid;
-        }
-    }
+    reducers
 });
 
 export const { toggleView } = viewSlice.actions;
