+++ my-app/src/state/store.js
@@ -1,8 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit';
 import playablePlaylistReducer from './slices/playablePlaylist/playablePlaylistSlice';
 import playerConfigReducer from './slices/playerConfig/playerConfigSlice';
-import filterReducer from './slices/filterSlice';
-import viewReducer from './slices/viewSlice';
+import filterReducer from './slices/filter/filterSlice';
+import viewReducer from './slices/view/viewSlice';
 
 export const store = configureStore({
   reducer: {
